# discord_bot_api/openapi.yaml
openapi: 3.0.3
info:
  title: Discord Authentication API
  description: HTTP API for interaction between the Discord bot and the Gothic 2 Online game server
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Authentication
    description: Discord authorization operations
  - name: Status
    description: API status check

paths:
  /discord/verify:
    post:
      tags:
        - Authentication
      summary: Verify authorization code
      description: Verifies the validity of the authorization code and returns session information
      operationId: verifyAuthCode
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - auth_code
              properties:
                auth_code:
                  type: string
                  description: Authorization code from the game (8 characters)
                  example: "ABCDEFGH"
                  minLength: 8
                  maxLength: 8
            examples:
              validCode:
                value:
                  auth_code: "ABCDEFGH"
      responses:
        '200':
          description: Code verification result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/VerifySuccessResponse'
                  - $ref: '#/components/schemas/VerifyErrorResponse'
              examples:
                valid:
                  value:
                    valid: true
                    player_id: 123
                    expires_at: 1704067200000
                invalid:
                  value:
                    valid: false
                    error: "INVALID_CODE"
                expired:
                  value:
                    valid: false
                    error: "CODE_EXPIRED"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /discord/link:
    post:
      tags:
        - Authentication
      summary: Link Discord ID to a game account
      description: Links a user's Discord ID to a game account using an authorization code
      operationId: linkDiscordAccount
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - discord_id
                - auth_code
              properties:
                discord_id:
                  type: string
                  description: User's Discord ID (Snowflake)
                  example: "123456789012345678"
                  pattern: '^\d+$'
                auth_code:
                  type: string
                  description: Authorization code from the game
                  example: "ABCDEFGH"
                  minLength: 8
                  maxLength: 8
            examples:
              validLink:
                value:
                  discord_id: "123456789012345678"
                  auth_code: "ABCDEFGH"
      responses:
        '200':
          description: Linking result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LinkSuccessResponse'
                  - $ref: '#/components/schemas/LinkErrorResponse'
              examples:
                success:
                  value:
                    success: true
                    player_id: 123
                invalidCode:
                  value:
                    success: false
                    error: "INVALID_CODE"
                expired:
                  value:
                    success: false
                    error: "CODE_EXPIRED"
                alreadyLinked:
                  value:
                    success: false
                    error: "DISCORD_ALREADY_LINKED"
                    message: "This Discord account is already linked to another game account"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /discord/status:
    get:
      tags:
        - Status
      summary: Check API status
      description: Returns the current status of the API server
      operationId: getStatus
      responses:
        '200':
          description: API status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: "ok"
                timestamp: "2024-01-01T12:00:00.000Z"
                version: "1.0.0"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authorizing requests

  schemas:
    VerifySuccessResponse:
      type: object
      required:
        - valid
        - player_id
        - expires_at
      properties:
        valid:
          type: boolean
          example: true
        player_id:
          type: integer
          description: Player ID in the database
          example: 123
        expires_at:
          type: integer
          description: Code expiration time in milliseconds (Unix timestamp)
          example: 1704067200000

    VerifyErrorResponse:
      type: object
      required:
        - valid
        - error
      properties:
        valid:
          type: boolean
          example: false
        error:
          type: string
          enum:
            - INVALID_CODE
            - CODE_EXPIRED
            - INTERNAL_ERROR
          description: Error code
          example: "INVALID_CODE"
        message:
          type: string
          description: Error description (optional)
          example: "auth_code is required"

    LinkSuccessResponse:
      type: object
      required:
        - success
        - player_id
      properties:
        success:
          type: boolean
          example: true
        player_id:
          type: integer
          description: Player ID in the database
          example: 123

    LinkErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          enum:
            - INVALID_CODE
            - CODE_EXPIRED
            - DISCORD_ALREADY_LINKED
            - PLAYER_NOT_FOUND
            - INTERNAL_ERROR
            - INVALID_REQUEST
          description: Error code
          example: "DISCORD_ALREADY_LINKED"
        message:
          type: string
          description: Error description
          example: "This Discord account is already linked to another game account"

    StatusResponse:
      type: object
      required:
        - status
        - timestamp
        - version
      properties:
        status:
          type: string
          example: "ok"
        timestamp:
          type: string
          format: date-time
          description: Current server time in ISO 8601 format
          example: "2024-01-01T12:00:00.000Z"
        version:
          type: string
          description: API version
          example: "1.0.0"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error code
          example: "INVALID_API_KEY"
        message:
          type: string
          description: Error description
          example: "Invalid API key"
